Added node_options environment to kibana
The Kibana server is routinely spking its memory usage and then getting
restarted by the OOM killer. In order to prvent this from happening it
is necessary to add a `NODE_OPTIONS` environment variable to pass a
`--max-old-space-size` flag to the node process.

Updated the state run to add this variable to the upstart or sytemd definitions.